collect very exact train & test data for learning digits
